refactor: move shared sandbox utils to src/lib and src/hooks
yaml-generator, json-response-parser (+ tests) and useMobile were physically
located under src/app/sandbox but imported by the live analysis viewer, the
story API routes, and the shared textarea UI primitive. Relocate them so the
sandbox route no longer owns cross-surface code.
